waiting_list_service/requirements.txt
fastapi==0.116.0
requests==2.32.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
uvicorn==0.35.0